requirements.txt
# =============================================================================
# Python Dependencies
# Install with: pip install -r requirements.txt
# Python version: 3.10+ (tested on 3.11, 3.12, 3.13, 3.14)
# =============================================================================

# Snowflake connectivity — base install only (no pandas extra)
# The [pandas] extra requires pandas, which has limited Python 3.14 support.
# Our pipeline does not use pandas — all data moves as Python dicts/JSON.
snowflake-connector-python==3.6.0

# HTTP requests to TMDb API
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# Retry logic with exponential backoff
tenacity==8.2.3

# Rich terminal output (logging, progress bars)
rich==13.7.0

# Type checking (development)
typing-extensions==4.9.0
